+++ scripts/codex/constants.js
@@ -50,6 +50,7 @@ export const FLAG_CHARACTER_CREATION = "greybeardCharacterCreation";
 export const FLAG_CREW_ROSTER = "crewRoster";
 export const FLAG_TRAINING_SKILL = "trainingSkill";
 export const FLAG_CYBERWARE = "cyberware";
+export const FLAG_SLICKWARE = "slickware";
 
 export const SETTING_ENABLE_CHARACTER_CREATOR = "enableCharacterCreator";
 export const SETTING_ENABLE_CYBERWARE = "enableCyberware";